+++ packages/react-magma-dom/src/i18n/interface.ts
@@ -141,6 +141,17 @@ export interface I18nInterface {
     menuItemSelectedAriaLabel: string;
     toggleMenuAriaLabel: string;
   };
+  fileUploader: {
+    browseFiles: string;
+    dragMessage: string;
+    errors: {
+      [key: string]: {
+        header?: string;
+        message: string;
+      };
+    }
+    files: string;
+  };
   header: {
     navigationButtonLabel: string;
     search: {