+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 2.2.6-next.2
+
+### Patch Changes
+
+- Updated dependencies [84507cc]
+  - react-magma-dom@2.5.5-next.2
+
+## 2.2.6-next.1
+
+### Patch Changes
+
+- Updated dependencies [ab506800]
+  - react-magma-dom@2.5.5-next.1
+
 ## 2.2.6-next.0
 
 ### Patch Changes